checkpoint it file format import support

+++ src/tracker/module_renderer.cpp
@@ -9,7 +9,7 @@ void module_renderer::change_audio_settings(const paudio_settings_t &settings) {
     auto set_deprecated_globals = [&settings] () {
         module_renderer::deprecated_global_bits_per_sample = 16;
         module_renderer::deprecated_global_channels = 2;
-        module_renderer::deprecated_global_mixing_freq = settings.sample_rate;
+        module_renderer::deprecated_global_mixing_freq = static_cast<uint32_t>(settings.sample_rate);
         module_renderer::deprecated_global_sound_setup_bitmask &= ~(SNDMIX_SOFTPANNING | SNDMIX_REVERSESTEREO);
     };
